# contracts/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Resilience API",
    "description": "Deterministic application boundary for the Resilience prototype.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/foundation/bootstrap": {
      "get": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Bootstrap",
        "operationId": "bootstrap_api_v1_foundation_bootstrap_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationBootstrap"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/foundation/onboarding": {
      "put": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Onboarding",
        "operationId": "onboarding_api_v1_foundation_onboarding_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OnboardingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationBootstrap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/foundation/profile": {
      "patch": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Profile Update",
        "operationId": "profile_update_api_v1_foundation_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/foundation/recurring-work-costs/{item_id}": {
      "put": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Recurring Put",
        "operationId": "recurring_put_api_v1_foundation_recurring_work_costs__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecurringWorkCostInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecurringWorkCostResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Recurring Delete",
        "operationId": "recurring_delete_api_v1_foundation_recurring_work_costs__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foundation/essential-expenses/{item_id}": {
      "put": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Essential Put",
        "operationId": "essential_put_api_v1_foundation_essential_expenses__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EssentialExpenseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EssentialExpenseResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Essential Delete",
        "operationId": "essential_delete_api_v1_foundation_essential_expenses__item_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foundation/weeks": {
      "get": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Weeks List",
        "operationId": "weeks_list_api_v1_foundation_weeks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WeeklyEntryResponse"
                  },
                  "title": "Response Weeks List Api V1 Foundation Weeks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foundation/weeks/{week_start}": {
      "get": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Week Get",
        "operationId": "week_get_api_v1_foundation_weeks__week_start__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "week_start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Week Start"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Week Put",
        "operationId": "week_put_api_v1_foundation_weeks__week_start__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "week_start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Week Start"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyEntryUpsert"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeeklyEntryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Week Delete",
        "operationId": "week_delete_api_v1_foundation_weeks__week_start__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "week_start",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Week Start"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foundation/transactions": {
      "post": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Transaction Create",
        "operationId": "transaction_create_api_v1_foundation_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/foundation/transactions/{transaction_id}": {
      "delete": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Transaction Delete",
        "operationId": "transaction_delete_api_v1_foundation_transactions__transaction_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/foundation/data": {
      "delete": {
        "tags": [
          "foundation-input"
        ],
        "summary": "Data Reset",
        "operationId": "data_reset_api_v1_foundation_data_delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "X-Confirm-Reset",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Confirm-Reset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoundationBootstrap"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EssentialExpenseInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "housing",
              "food",
              "transport",
              "utilities",
              "healthcare",
              "caregiving",
              "debt",
              "other"
            ],
            "title": "Category"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label"
          },
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "cadence": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly"
            ],
            "title": "Cadence"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "label",
          "amountCents",
          "cadence"
        ],
        "title": "EssentialExpenseInput"
      },
      "EssentialExpenseResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "housing",
              "food",
              "transport",
              "utilities",
              "healthcare",
              "caregiving",
              "debt",
              "other"
            ],
            "title": "Category"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label"
          },
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "cadence": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly"
            ],
            "title": "Cadence"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "label",
          "amountCents",
          "cadence",
          "createdAt",
          "updatedAt"
        ],
        "title": "EssentialExpenseResponse"
      },
...
      "FoundationBootstrap": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/ProfileResponse"
          },
          "recurringWorkCosts": {
            "items": {
              "$ref": "#/components/schemas/RecurringWorkCostResponse"
            },
            "type": "array",
            "title": "Recurringworkcosts"
          },
          "essentialExpenses": {
            "items": {
              "$ref": "#/components/schemas/EssentialExpenseResponse"
            },
            "type": "array",
            "title": "Essentialexpenses"
          },
          "weeklyEntries": {
            "items": {
              "$ref": "#/components/schemas/WeeklyEntryResponse"
            },
            "type": "array",
            "title": "Weeklyentries"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionResponse"
            },
            "type": "array",
            "title": "Transactions"
          },
          "syncedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Syncedat"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "recurringWorkCosts",
          "essentialExpenses",
          "weeklyEntries",
          "syncedAt"
        ],
        "title": "FoundationBootstrap"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OnboardingRequest": {
        "properties": {
          "emergencySavingsCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Emergencysavingscents"
          },
          "recurringWorkCosts": {
            "items": {
              "$ref": "#/components/schemas/RecurringWorkCostInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Recurringworkcosts"
          },
          "essentialExpenses": {
            "items": {
              "$ref": "#/components/schemas/EssentialExpenseInput"
            },
            "type": "array",
            "maxItems": 30,
            "title": "Essentialexpenses"
          },
          "firstWeekStart": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Firstweekstart"
          },
          "firstWeek": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WeeklyEntryUpsert"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "emergencySavingsCents"
        ],
        "title": "OnboardingRequest"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "onboardingCompleted": {
            "type": "boolean",
            "title": "Onboardingcompleted"
          },
          "latestEmergencySavingsCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "minimum": 0.0,
            "title": "Latestemergencysavingscents"
          },
          "emergencyFundBalanceCents": {
            "type": "integer",
            "title": "Emergencyfundbalancecents",
            "default": 0
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "phoneNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonenumber"
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateofbirth"
          }
        },
        "type": "object",
        "required": [
          "id",
          "currency",
          "timezone",
          "onboardingCompleted",
          "latestEmergencySavingsCents"
        ],
        "title": "ProfileResponse"
      },
      "ProfileUpdate": {
        "properties": {
          "displayName": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "phoneNumber": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Phonenumber"
          },
          "dateOfBirth": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dateofbirth"
          }
        },
        "type": "object",
        "title": "ProfileUpdate"
      },
...
      "RecurringWorkCostInput": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "vehicle_rental",
              "insurance",
              "subscription",
              "equipment",
              "other"
            ],
            "title": "Category"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label"
          },
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "cadence": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly"
            ],
            "title": "Cadence"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "label",
          "amountCents",
          "cadence"
        ],
        "title": "RecurringWorkCostInput"
      },
      "RecurringWorkCostResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "category": {
            "type": "string",
            "enum": [
              "vehicle_rental",
              "insurance",
              "subscription",
              "equipment",
              "other"
            ],
            "title": "Category"
          },
          "label": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Label"
          },
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "cadence": {
            "type": "string",
            "enum": [
              "weekly",
              "monthly"
            ],
            "title": "Cadence"
          },
          "isActive": {
            "type": "boolean",
            "title": "Isactive",
            "default": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "category",
          "label",
          "amountCents",
          "cadence",
          "createdAt",
          "updatedAt"
        ],
        "title": "RecurringWorkCostResponse"
      },
...
      "TransactionInput": {
        "properties": {
          "entryType": {
            "type": "string",
            "enum": [
              "income",
              "cost"
            ],
            "title": "Entrytype"
          },
          "amountCents": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Amountcents"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 160
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "occurredOn": {
            "type": "string",
            "format": "date",
            "title": "Occurredon"
          }
        },
        "type": "object",
        "required": [
          "entryType",
          "amountCents",
          "occurredOn"
        ],
        "title": "TransactionInput"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}